requirements.txt
asgiref==3.8.1
attrs==25.1.0
cfgv==3.4.0
coverage==7.6.12
distlib==0.3.9
Django==5.1.6
django-extensions==3.2.3
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
Faker==36.1.1
filelock==3.17.0
identify==2.6.7
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
nodeenv==1.9.1
packaging==24.2
platformdirs==4.3.6
pluggy==1.5.0
psycopg2==2.9.10
PyJWT==2.10.1
pytest==8.3.4
pytest-cov==6.0.0
pytest-django==4.10.0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.23.1
ruff==0.9.7
sqlparse==0.5.3
tzdata==2025.1
uritemplate==4.1.1
virtualenv==20.29.2
